server/requirements.txt
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.39.13
botocore==1.39.13
click==8.1.8
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.116.0
future==0.18.2
h11==0.16.0
idna==3.10
jmespath==1.0.1
macholib==1.15.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
rsa==4.9.1
s3transfer==0.13.1
six==1.15.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
